+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This file contains highlights of what changes on each version of the [Grinder-PH
 
 ## Version [1.0.0](https://github.com/cedx/grinder-php-minify/compare/v0.2.0...v1.0.0)
 - Breaking change: changed the signature of most class constructors.
+- Breaking change: merged the `compress()` and `compressFile()` functions into the `phpMinify()` function.
 - Breaking change: raised the required [Dart](https://www.dartlang.org) version.
 - Breaking change: removed the `binary` and `mode` properties from the `Minifier` class.
 - Breaking change: the `Minifier.transformer` property is now `final`.